Audio now working
With this, despite some errata, all major components of the board are verified.

The audio is working, despite some misunderstandings here and there. Unfortunately the audio is horribly loud also, so the final board will hopefully address some improvements that can be made there.

Had to change TCA0 to Split, as misunderstandings during the board layout forced me to use split mode in order to get an output.