Refactored statemachine/ ISRs to work with sleep
This commit changes the "board_state_waiting" state such that instead of endlessly waiting for enough consecutive PIR detects to enter countdown mode, it now is prepared to fall down to a atTINY 1606 sleep mode (exact sleep mode to be determined). Hysteresis is also added to this "fall to sleep" counter so that the board does not thrash too much between low power mode and wakeup during intermittent detection.

The board appears to behave nominally with these changes.

NOTE: for the moment the sleep command is disabled, though the preparations have mostly been made for it to be enabled. .