Improve generated Doxygen
Changes to documentation to improve the output Doxygen, seeing as anyways the code has changed and thus the Doxygen must as well. The changes to the Doxyfile are fairly significant and resolve some long-standing frustrations I'd had with the layout of the resulting HTML, especially with regards to finding the "Files" pages.