Update win/lose audio files
At long last the win and lose audio files make sense and in fact tell you if you win or lose.

Although the code changes are mostly just replacing a struct, a lot of interesting work occurred out-of-band in order to make the phrases "you win" and "you lose" intelligible and able to fit in 16kB of flash.

I intend to write it up in a README in the next commit in this PR.