Add Powershell Scripts used during development
These are scripts that I used and would like to keep around if this project is revisited in the future. It is trivial to imagine converting them to bash or whatever shell- the key is just having pymcuprog in the path and the rest is trivial. The main scripts used here are `EraseATTiny.ps1` which I used before every re-flash of the board during development, `ProgramATTIny.ps1` which I didn't actually use directly but instead fed into a build configuration of Microchip Studio and used as my "Start Without Debugging" command, and `WriteATTinyResetFuse_DESTRUCTIVE.ps1` which I used at the end of the project to enable the reset button by reconfiguring the RESET/UPDI pin to do reset instead of UPDI